% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slash.R
\name{slasher}
\alias{slasher}
\alias{fslash}
\alias{bslash}
\title{Edit the slash type of the path that is currently in the clipboard}
\usage{
fslash()

bslash()
}
\description{
\code{fslash} converts slashes in the clipboard to forward slashes;
\code{bslash} converts slashses in the clipboard to back slashes.

Both of these functions are designed to be used while copying and pasting
paths between R and other applications on a Windows system. For example you
could copy a path from Windows Explorer, run \code{fslash()} and then paste
the modfied path into R. Conversely you could copy a path from your R code
run \code{bslash()} and then paste into Explorer or other applications that
expect a backslash.
}
