% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winopen.R
\name{winopen}
\alias{winopen}
\title{Function to open a file or folder}
\usage{
winopen(x)
}
\arguments{
\item{x}{A folder or file to be opened.}
}
\value{
This function is primarily called for the side effect of opening a
file or folder. It invisibly returns a normalized path to the item it
opened, which is either \code{x} or a parent directory that exists on disk.
}
\description{
This function will open the target path, \code{x},  with it's associated
application, or with Windows Explorer if \code{x} is a folder
}
\details{
If the file or folder does not exist it will open the first parent
folder that does exists.
}
\examples{
\dontrun{

# Open the R directory
winopen("C:/Program Files/R/")

# Plot to a file and then open the file
f <- tempfile(fileext = ".png")
png(f)
plot(1, 1)
dev.off()
winopen(f)
}
}
